Query generator v4 - Query Subsampling (#307)
## What is the goal of this PR?
We not only want to generate queries, we want to generate a wide variety of queries. The work so far has been progressing towards this already, providing a vector quantification of queries in the prior PRs. This PR implements two different ways of filtering a large number of queries into a much smaller set.

Approach 1 - K-means:
Cluster the queries into a number of queries (as given) and choose some number of queries from each cluster. 
Advantages: visual inspection shows that a similar queries are grouped together successfully. 
Disadvantages: Difficult to visaulise/consider what the covered space actually is or means, finding a good cluster initialisation is difficult - with more than 30 target clusters we often end up with 30 filled clusters and the remainder empty, which have to be eliminated. Thus, it is hard to generate more than 30 distinct "types" of queries

Approach 2 - Gridded:
Take each axis in the vector space that represents the query, and divide it into 5 (or some other number if desired). This generates a coordinate for each "cube"/"grid square" that each query lives in. We then choose one query from each grid square that we know is populated.

Advantages: well defined what it means to have good coverage of the space - ideally one sample from each grid square (though this is very high dimensional, like 10^5). We could consider feeding back into the query generation to generates types of queries not seen before. 
Disadvantages; can be very granular and may end up letting us keep quite similar queries. Ideal coverage is very large - 10^5 as stated above.

Finishes #300 

## What are the changes implemented in this PR?
* `query-sampler-binary` entry point for running the query sampler against a keyspace and producing some number of queries
* Implement KMeans and Gridded sampling algorithms to subsample queries
* Bump `Grakn` and `client_java` dependencies
* fix attribute comparison bug - now allow generating multiple comparisons between a variable and other variables
* new classes for wrapping a `GraqlGet` query with its vector representation
* Update to be compatible with new client-java